% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advancedPBRTQC.R
\name{makeDeviceMeanDiff}
\alias{makeDeviceMeanDiff}
\title{Factory for device-difference PBRTQC (mean)}
\usage{
makeDeviceMeanDiff()
}
\value{
A list with \code{fn} and \code{getStore} (same interface as
\code{\link{makeRegAdjEMA}})
}
\description{
At each time point computes the difference between the per-device rolling
mean and the overall rolling mean.  Requires a column \code{device} in
\code{dataExtra}.
}
