% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advancedPBRTQC.R
\name{regAdjEMA}
\alias{regAdjEMA}
\title{Regression adjusted PBRTQC}
\usage{
regAdjEMA(measurement, blockSize, ll, ul, dataExtra)
}
\arguments{
\item{measurement}{a vector of measurements}

\item{blockSize}{the block size}

\item{ll}{lower limit for winsorizing}

\item{ul}{upper limit for winsorizing}

\item{dataExtra}{}
}
\description{
Regression adjusted PBRTQC
}
