% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simPBRTQC}
\alias{simPBRTQC}
\title{Simualtion of PBRTQC methods}
\usage{
simPBRTQC(
  data,
  blockSizes,
  truncationLimits,
  biases,
  fxs,
  percAccAlarms,
  calcContinous,
  max_samples = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{blockSizes}{numeric vector}

\item{truncationLimits}{data.frame with columns 'lowerTrunc', 'upperTrunc'
and 'truncation'.}

\item{biases}{numeric vector}

\item{fxs}{named list of functions for PBRTQC algorithms.}

\item{percAccAlarms}{percentage of acceptable false alarms per day}

\item{max_samples}{optional integer. If set, at most this many samples per
day are used. If a \code{device} column is present, the limit applies per
device per day.}
}
\value{
data.frame with results
}
\description{
Simualtion of PBRTQC methods
}
\examples{
\dontrun{
biases <- bias * seq(-1.5, 1.5, by=0.2)
blockSizes <- seq(10, 150, by = 20)
fxs <- list("mean" = rollMean,
            'median' = rollMed,
            'EMA' = truncatedEMA)
 truncationLimits <- data.frame('lowerTrunc' = ll,
                                'upperTrunc' = ul,
                                'truncation' = 'userSelected')
 res <- simPBRTQC(dataForSim,
                  blockSizes, truncationLimits, biases, fxs,
                   0.05)
}


}
