% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisPBRTQC.R
\name{truncatedEMA}
\alias{truncatedEMA}
\title{Exponential Moving Average with winsorizing}
\usage{
truncatedEMA(measurement, blockSize, ll, ul)
}
\arguments{
\item{measurement}{a vector of measurements}

\item{blockSize}{the block size}

\item{ll}{lower limit for winsorizing}

\item{ul}{upper limit for winsorizing}
}
\value{
vector of exponential moving averages
}
\description{
Exponential Moving Average with winsorizing
}
